+++ src/main/java/dev/nicho/rolesync/db/SQLiteHandler.java
@@ -1,6 +1,6 @@
 package dev.nicho.rolesync.db;
 
-import org.bukkit.plugin.java.JavaPlugin;
+import dev.nicho.rolesync.RoleSync;
 
 import java.io.File;
 import java.io.IOException;
@@ -13,7 +13,7 @@ public class SQLiteHandler extends DatabaseHandler {
     private final File db;
     private Connection connection;
 
-    public SQLiteHandler(JavaPlugin plugin, File db) throws IOException, SQLException {
+    public SQLiteHandler(RoleSync plugin, File db) throws IOException, SQLException {
         super(plugin);
 
         if (!db.exists()) {